Updated main project with NTP time

+++ arduino/Report.cpp
@@ -1,10 +1,10 @@
 #include "Report.h"
 
-Report::Report(double unixtime, int sensorcount)
+Report::Report(uint32_t unixtime, int sensorcount)
 	: jsonBuffer(JSON_ARRAY_SIZE(sensorcount) + JSON_OBJECT_SIZE(2) + sensorcount * JSON_OBJECT_SIZE(4))
 {
 	root= &jsonBuffer.createObject();
-	root->set("time", unixtime);
+	root->set("time",unixtime);
 	measurements = &root->createNestedArray("sensor");
 }
